chore: move context type to lib/apollo
