+++ src/graphql/__mocks__/apolloServer.ts
@@ -3,9 +3,8 @@ import { FastifySessionObject } from "@fastify/session";
 import { FastifyReply, FastifyRequest } from "fastify";
 import { mock, mockDeep } from "jest-mock-extended";
 
-import { formatError } from "@/lib/apolloServer.js";
+import { formatError, IContext } from "@/lib/apolloServer.js";
 
-import { IContext } from "../context.js";
 import { resolvers } from "../resolvers.js";
 import typeDefs from "../type-defs.js";
