--- src/graphql/index.ts
@@ -1,2 +1,0 @@
-export { resolvers } from "./resolvers.js";
-export { typeDefs } from "./type-defs.js";
